Bot joins and no errors
No sound is played when bot joins the voice channel. This branch might need to be revised and looked into if its actually possible.

Its also maybe to do with youtube not wanting this use of their videos so there are guards against this behavior, i have not looked into that so i will pause this branch as of right now.